requirements.txt
affine==2.3.0
asgiref==3.3.4
attrs==21.2.0
boto3==1.18.54
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click-plugins==1.1.1
cligj==0.7.2
django==3.2.6
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
geoip2==4.5.0
geopandas==0.9.0
h5py==3.3.0
idna==2.10
pandas==1.2.5
psycopg2==2.9.1
pyproj==3.1.0
pytz==2021.1
rasterio==1.2.6
requests==2.25.1
rioxarray==0.6.1
scipy==1.7.0
shapely==1.7.1
snuggs==1.4.7
urllib3==1.26.6
wget==3.2